Unify code (shared library)
This unifies the code between extensions, [using the standard APIs first](https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API#browser_api_differences), removing a lot of duplicate code and unnecessary as well (like different ways of setting the `Authorization` header), as well as other minor tweaks/refactoring for simpler understanding and better readability.

I've tested it extensively in Chrome and Firefox, but would like to have others test in Edge/Brave/Opera to see if all is working as expected.

This is the final DX improvement to be made before some other UX improvements will be worked on, as it makes updating and improving extensions much simpler. Basically the only _file_ differences now are in the `manifest.json`.
