+++ chrome/manifest.json
@@ -4,7 +4,8 @@
   "version": "0.3",
   "description": "A simple helper extension for setting Kagi as a default search engine, and automatically logging in to Kagi in incognito browsing windows",
   "background": {
-    "service_worker": "src/background.js"
+    "service_worker": "src/background.js",
+    "type": "module"
   },
   "icons": {
     "16": "icons/icon_16px.png",
