+++ package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "description": "A simple helper extension for setting Kagi as a default search engine, and automatically logging in to Kagi in incognito browsing windows",
   "main": "build.js",
+  "private": true,
   "scripts": {
     "watch-firefox": "nodemon build.js firefox watch",
     "watch-chrome": "nodemon build.js chrome watch",
